# docs/api/openapi.json
{
  "components": {
    "schemas": {
...
      "ConnectionsResponse": {
        "description": "Envelope for the connection list.\n\nA wrapper object rather than a bare array leaves room for pagination or\nmetadata later without breaking the generated Swift client.\n\nAttributes\n----------\nconnections : list[ConnectionResponse]\n    The caller's connections, oldest first.",
        "properties": {
          "connections": {
            "items": {
              "$ref": "#/components/schemas/ConnectionResponse"
            },
            "title": "Connections",
            "type": "array"
          }
        },
        "required": [
          "connections"
        ],
        "title": "ConnectionsResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "StartConnectionRequest": {
        "description": "Request to begin authorizing a bank connection.\n\nAttributes\n----------\ninstitution : str\n    The bank's provider-scoped identifier (Enable Banking ASPSP ``name``).\ncountry : str\n    ISO 3166-1 alpha-2 country of the bank.\nlogo : str or None\n    The bank's logo URL, taken from the same ``GET /connections/institutions``\n    entry the caller picked. Stored verbatim on the connection for the\n    Conti screen; ``None`` when the picker had none. Cosmetic \u2014 the client\n    already had it, so re-fetching it server-side would only add a\n    provider round-trip to the SCA-start path.",
        "properties": {
          "country": {
            "title": "Country",
            "type": "string"
          },
          "institution": {
            "title": "Institution",
            "type": "string"
          },
          "logo": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Logo"
          }
        },
        "required": [
          "institution",
          "country"
        ],
        "title": "StartConnectionRequest",
        "type": "object"
      },
      "StartConnectionResponse": {
        "description": "The started authorization: where to send the user.\n\nAttributes\n----------\nconnection_id : UUID\n    The pending connection created for this authorization.\nauthorization_url : str\n    The bank authorization URL to open in the system browser.",
        "properties": {
          "authorization_url": {
            "title": "Authorization Url",
            "type": "string"
          },
          "connection_id": {
            "format": "uuid",
            "title": "Connection Id",
            "type": "string"
          }
        },
        "required": [
          "connection_id",
          "authorization_url"
        ],
        "title": "StartConnectionResponse",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Traccio",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/connections": {
      "get": {
        "description": "List the current user's bank connections, oldest first.\n\nScoped to the current user. Secret material never leaves ``db/``, so the\nprojection cannot expose the consent secret or ``auth_state`` (see\n``docs/engineering.md``).\n\nParameters\n----------\nsession : Session\n    Request-scoped database session.\nuser_id : UUID\n    The user whose connections to return.\n\nReturns\n-------\nConnectionsResponse\n    The user's connections, oldest first.",
        "operationId": "connections_connections_get",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConnectionsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Connections"
      },
      "post": {
        "description": "Begin authorizing a bank connection.\n\nCreates a pending connection and returns the URL to open in the system\nbrowser for SCA. Scoped to the current user.\n\nParameters\n----------\nbody : StartConnectionRequest\n    The bank to authorize (institution + country, plus its logo URL from\n    the picker, stored verbatim for display).\nsession : Session\n    Request-scoped database session.\nuser_id : UUID\n    The user the connection belongs to.\nprovider : BankProvider\n    The bank adapter (Enable Banking).\n\nReturns\n-------\nStartConnectionResponse\n    The pending connection id and the authorization url.",
        "operationId": "start_connection_connections_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StartConnectionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StartConnectionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Start Connection"
      }
    },
    "/connections/backfill-logos": {
      "post": {
        "description": "Fill in ``institution_logo`` for connections that predate it.\n\n``institution_logo`` is written only when a connection is first created\n(``POST /connections``); connections authorized before that column existed\n(migration ``e5f6a7b8c9d0``) carry ``NULL`` and show a lettermark in the\nclient. This looks each one up in the provider's institution list for its\ncountry, matches by ``institution_name`` (case- and whitespace-insensitive),\nand writes the logo URL the provider reports.\n\nA connection that predates the ``country`` column (migration\n``b8f3d2e7c1a4``) also carries ``NULL`` there \u2014 the same rows, since both\ncolumns were added within a week of each other. Those fall back to\n:attr:`Settings.default_institution_country`, and the resolved country is\npersisted alongside the logo, which additionally unblocks in-place\nre-authorization for them (``POST /connections/{id}/reauthorize`` ``409``s\non a ``NULL`` country).\n\nIdempotent and safe to call repeatedly: a connection that already has a\nlogo or matches no provider institution is left untouched. One provider\nlookup per distinct country, not per connection.\n\nParameters\n----------\nsession : Session\n    Request-scoped database session.\nuser_id : UUID\n    The user whose connections to backfill.\nprovider : BankProvider\n    The bank adapter (Enable Banking).\n\nReturns\n-------\nBackfillLogosResponse\n    How many connections gained a logo this call.\n\nRaises\n------\nHTTPException\n    502 if a provider institution lookup fails.",
        "operationId": "backfill_connection_logos_connections_backfill_logos_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BackfillLogosResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Backfill Connection Logos"
      }
    },
    "/connections/callback": {
      "get": {
        "description": "Complete an authorization after the bank's SCA redirect.\n\nMatches the callback to its pending connection by ``state``, exchanges the\ncode for a session, encrypts the consent secret at rest, and activates the\nconnection. Scoped to the current user.\n\nParameters\n----------\nsession : Session\n    Request-scoped database session.\nuser_id : UUID\n    The user the connection belongs to.\nprovider : BankProvider\n    The bank adapter (Enable Banking).\ncipher : Fernet cipher\n    Encrypts the consent secret before it is stored.\ncode, state, error, error_description : str or None\n    The query parameters the bank appended to the redirect.\n\nReturns\n-------\nHTMLResponse\n    A small success page for the browser tab.",
        "operationId": "connection_callback_connections_callback_get",
        "parameters": [
          {
            "in": "query",
            "name": "code",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Code"
            }
          },
          {
            "in": "query",
            "name": "state",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "State"
            }
          },
          {
            "in": "query",
            "name": "error",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Error"
            }
          },
          {
            "in": "query",
            "name": "error_description",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Error Description"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "text/html": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Connection Callback"
      }
    },
...
  }
}